UI: Extracted API access layer
